+++ guice/src/main/java/org/jdbi/v3/guice/internal/InternalLooseImportBindingBuilder.java
@@ -17,11 +17,11 @@ import java.lang.annotation.Annotation;
 
 import javax.annotation.CheckForNull;
 import javax.annotation.Nullable;
-import javax.inject.Inject;
-import javax.inject.Provider;
 
+import com.google.inject.Inject;
 import com.google.inject.Injector;
 import com.google.inject.Key;
+import com.google.inject.Provider;
 import com.google.inject.Scope;
 import com.google.inject.TypeLiteral;
 import com.google.inject.binder.LinkedBindingBuilder;