+++ guice/src/test/java/org/jdbi/v3/guice/TestJdbiBinder.java
@@ -21,10 +21,9 @@ import java.time.OffsetDateTime;
 import java.util.List;
 import java.util.UUID;
 
-import javax.inject.Inject;
-import javax.inject.Qualifier;
 import javax.sql.DataSource;
 
+import com.google.inject.Inject;
 import com.google.inject.Injector;
 import com.google.inject.Module;
 import com.google.inject.Scopes;
@@ -131,7 +130,8 @@ public class TestJdbiBinder {
 
     @Retention(RUNTIME)
     @Target({FIELD, PARAMETER, METHOD})
-    @Qualifier
+    @javax.inject.Qualifier
+    @jakarta.inject.Qualifier
     public @interface BinderTest {}
 
     public interface Dao {