+++ RELEASE_NOTES.md
@@ -1,5 +1,6 @@
 # Unreleased
   - Handle.inTransaction: improve exception thrown when restoring transaction isolation #2343
+  - add support for Guice 6.x (using javax.inject annotations) and guice 7.x (using jakarta.inject annotations)
 
 # 3.38.2
   - spring5 JdbiUtil: fix thread safety #2341